pyproject.toml
[project]
name = "reddit-mcp"
version = "0.2.0"
description = "MCP server for searching and browsing Reddit posts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]==1.27.0",
    "asyncpraw==7.8.1",
    "python-dotenv==1.2.2",
    "cachetools==7.0.5",
]

[dependency-groups]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
]

[project.scripts]
reddit-mcp = "reddit_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reddit_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
